move routeoptions extension plugin to correct package structure
