+++ projects/gateway2/translator/extensions/extensions_plugin_registry.go
@@ -1,31 +1,41 @@
 package extensions
 
 import (
+	"context"
 	"fmt"
 
-	v1 "github.com/solo-io/gloo/projects/gateway/pkg/api/v1"
-	"github.com/solo-io/gloo/projects/gateway2/translator/httproute/filterplugins"
-	"github.com/solo-io/gloo/projects/gateway2/translator/httproute/filterplugins/routeoptions"
+	sologatewayv1 "github.com/solo-io/gloo/projects/gateway/pkg/api/v1"
+	"github.com/solo-io/gloo/projects/gateway2/translator/extensions/routeoptions"
+	sologloov1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
 
 	"k8s.io/apimachinery/pkg/runtime/schema"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
+type ExtensionPlugin interface {
+	ApplyExtPlugin(
+		ctx context.Context,
+		cfg client.Object,
+		outputRoute *sologloov1.Route,
+	) error
+}
+
 type ExtensionPluginRegistry struct {
-	extensionPlugins map[schema.GroupKind]filterplugins.ExtensionPlugin
+	extensionPlugins map[schema.GroupKind]ExtensionPlugin
 }
 
 func NewExtensionPluginRegistry() *ExtensionPluginRegistry {
 	return &ExtensionPluginRegistry{
-		extensionPlugins: map[schema.GroupKind]filterplugins.ExtensionPlugin{
+		extensionPlugins: map[schema.GroupKind]ExtensionPlugin{
 			{
-				Group: v1.RouteOptionGVK.Group,
-				Kind:  v1.RouteOptionGVK.Kind,
+				Group: sologatewayv1.RouteOptionGVK.Group,
+				Kind:  sologatewayv1.RouteOptionGVK.Kind,
 			}: routeoptions.NewPlugin(),
 		},
 	}
 }
 
-func (h *ExtensionPluginRegistry) GetExtensionPlugin(gk schema.GroupKind) (filterplugins.ExtensionPlugin, error) {
+func (h *ExtensionPluginRegistry) GetExtensionPlugin(gk schema.GroupKind) (ExtensionPlugin, error) {
 	p, ok := h.extensionPlugins[schema.GroupKind{
 		Group: string(gk.Group),
 		Kind:  string(gk.Kind),
