+++ projects/gateway2/translator/extensions/routeoptions/route_options_plugin.go
@@ -1,9 +1,10 @@
 package routeoptions
 
 import (
+	"context"
+
 	"github.com/rotisserie/eris"
 	solokubev1 "github.com/solo-io/gloo/projects/gateway/pkg/api/v1/kube/apis/gateway.solo.io/v1"
-	"github.com/solo-io/gloo/projects/gateway2/translator/httproute/filterplugins"
 	v1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
 	"github.com/solo-io/go-utils/contextutils"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -18,13 +19,13 @@ func NewPlugin() *Plugin {
 }
 
 func (p *Plugin) ApplyExtPlugin(
-	ctx *filterplugins.RouteContext,
+	ctx context.Context,
 	cfg client.Object,
 	outputRoute *v1.Route,
 ) error {
 	routeOption, ok := cfg.(*solokubev1.RouteOption)
 	if !ok {
-		contextutils.LoggerFrom(ctx.Ctx).DPanic(incorrectTypeMsg)
+		contextutils.LoggerFrom(ctx).DPanic(incorrectTypeMsg)
 		return eris.Errorf(incorrectTypeMsg)
 	}
