+++ projects/gateway2/translator/httproute/filterplugins/filtertests/filter_plugin_test_utils.go
@@ -11,21 +11,8 @@ import (
 	"github.com/solo-io/gloo/projects/gateway2/translator/testutils"
 
 	v1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 	gwv1 "sigs.k8s.io/gateway-api/apis/v1"
 )
-
-func AssertExpectedRouteExtPlugin(
-	plugin filterplugins.ExtensionPlugin,
-	cfg client.Object,
-	expectedRoute *v1.Route,
-	logActual bool,
-) {
-	outputRoute := &v1.Route{
-		Options: &v1.RouteOptions{},
-	}
-	assertExpectedRouteExt(plugin, cfg, outputRoute, expectedRoute, nil, logActual)
-}
 
 func AssertExpectedRoute(
 	plugin filterplugins.FilterPlugin,
@@ -50,36 +37,6 @@ func AssertExpectedRouteWith(
 	assertExpectedRoute(plugin, filter, outputRoute, expectedRoute, match, logActual)
 }
 
-func assertExpectedRouteExt(
-	plugin filterplugins.ExtensionPlugin,
-	cfg client.Object,
-	outputRoute *v1.Route,
-	expectedRoute *v1.Route,
-	match *gwv1.HTTPRouteMatch,
-	logActual bool,
-) {
-	ctx := &filterplugins.RouteContext{
-		Ctx:      context.TODO(),
-		Route:    &gwv1.HTTPRoute{},
-		Queries:  nil,
-		Rule:     nil,
-		Reporter: nil,
-		Match:    match,
-	}
-	err := applyExtPlugin(plugin, ctx, cfg, outputRoute)
-	Expect(err).NotTo(HaveOccurred())
-
-	if logActual {
-		actualYaml, err := testutils.MarshalYaml(outputRoute)
-		Expect(err).NotTo(HaveOccurred())
-		log.Print("actualYaml: \n---\n", string(actualYaml), "\n---\n")
-		expectedYaml, err := testutils.MarshalYaml(expectedRoute)
-		Expect(err).NotTo(HaveOccurred())
-		log.Print("expectedYaml: \n---\n", string(expectedYaml), "\n---\n")
-	}
-	Expect(proto.Equal(outputRoute, expectedRoute)).To(BeTrue())
-}
-
 func assertExpectedRoute(
 	plugin filterplugins.FilterPlugin,
 	filter gwv1.HTTPRouteFilter,
@@ -96,7 +53,11 @@ func assertExpectedRoute(
 		Reporter: nil,
 		Match:    match,
 	}
-	err := applyPlugin(plugin, ctx, filter, outputRoute)
+	err := plugin.ApplyFilter(
+		ctx,
+		filter,
+		outputRoute,
+	)
 	Expect(err).NotTo(HaveOccurred())
 
 	if logActual {
@@ -108,30 +69,4 @@ func assertExpectedRoute(
 		log.Print("expectedYaml: \n---\n", string(expectedYaml), "\n---\n")
 	}
 	Expect(proto.Equal(outputRoute, expectedRoute)).To(BeTrue())
-}
-
-func applyExtPlugin(
-	plugin filterplugins.ExtensionPlugin,
-	ctx *filterplugins.RouteContext,
-	cfg client.Object,
-	outputRoute *v1.Route,
-) error {
-	return plugin.ApplyExtPlugin(
-		ctx,
-		cfg,
-		outputRoute,
-	)
-}
-
-func applyPlugin(
-	plugin filterplugins.FilterPlugin,
-	ctx *filterplugins.RouteContext,
-	filter gwv1.HTTPRouteFilter,
-	outputRoute *v1.Route,
-) error {
-	return plugin.ApplyFilter(
-		ctx,
-		filter,
-		outputRoute,
-	)
 }
